backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
openai>=0.28.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.7
pydantic>=2.3.0
aiohttp>=3.9.0
tenacity>=8.2.0
slowapi>=0.1.9
